Use a value set on the state itself rather than an isReady flag
When redux initializes it passes a couple of actions to each reducer.
This initialization was setting the "isReady" flag. If we have an
initial state coming from outside of the reducer, however, then on the
actual @@INIT pass the isReady flag would be true and the state be
defined even though it isn't actually initialized.

By using a flag on the state itself we can ensure that the state is
properly initialized.

Fixes #16